tests/wpt/web-platform-tests/wake-lock/wakelock-abortsignal-set.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WakeLock: passing an AbortSignal already set aborts</title>
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(t => {
  const abortController = new AbortController();
  const abortSignal = abortController.signal;
  abortController.abort();
  assert_true(abortSignal.aborted);

  return promise_rejects(t, "AbortError", WakeLock.request('screen', { signal: abortSignal }));
}, "A WakeLock request with an AbortSignal whose abort flag is set always aborts");

promise_test(async t => {
  const abortController = new AbortController();
  const abortSignal = abortController.signal;
  abortController.abort();
  assert_true(abortSignal.aborted);

  const lock1 = WakeLock.request('screen', { signal: abortSignal });
  const lock2 = WakeLock.request('screen', { signal: abortSignal });
  const lock3 = WakeLock.request('system', { signal: abortSignal });

  await promise_rejects(t, "AbortError", lock1);
  await promise_rejects(t, "AbortError", lock2);
  await promise_rejects(t, "AbortError", lock3);
}, "The same AbortSignal can be used to cause multiple wake locks to abort");
</script>
